# docs/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Movie API",
    "description": "A simple API for managing movie data, integrated with a MySQL database using Knex.js. This API allows listing, searching, adding, updating, and deleting movie records, and includes basic user authentication.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List all users",
        "description": "Retrieves a list of all registered users. (Note: In a production app, this would typically be an admin-only route.)",
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                },
                "examples": {
                  "success": {
                    "summary": "Example User List",
                    "value": [
                      { "id": 1, "email": "user1@example.com" },
                      { "id": 2, "email": "user2@example.com" }
                    ]
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/users/register": {
      "post": {
        "tags": ["Users"],
        "summary": "Register a new user",
        "description": "Registers a new user with an email and password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterInput"
              },
              "examples": {
                "newUser": {
                  "summary": "Example New User Registration",
                  "value": {
                    "email": "test@example.com",
                    "password": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "New User added successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer", "example": 3 },
                        "email": { "type": "string", "example": "test@example.com" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing email or password",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Request body incomplete, both email and password are required" }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict - User already exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "User already exists - please user different email." }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Login a user",
        "description": "Authenticates a user and issues a JWT token as a cookie.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginInput"
              },
              "examples": {
                "userLogin": {
                  "summary": "Example User Login",
                  "value": {
                    "email": "test@example.com",
                    "password": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, JWT token issued",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string",
                  "example": "token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; Path=/; Expires=Wed, 03 Jul 2025 00:00:00 GMT; HttpOnly"
                },
                "description": "Authentication token set as an HttpOnly cookie"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing email or password",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Email and password credentials inccorect or not provided." }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Invalid credentials" }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "description": "User profile details",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "email": { "type": "string", "format": "email", "example": "user@example.com" }
        },
        "required": ["id", "email"]
      },
      "UserRegisterInput": {
        "type": "object",
        "description": "Input for user registration",
        "properties": {
          "email": { "type": "string", "format": "email", "example": "newuser@example.com" },
          "password": { "type": "string", "format": "password", "example": "securepassword123" }
        },
        "required": ["email", "password"]
      },
      "UserLoginInput": {
        "type": "object",
        "description": "Input for user login",
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "format": "password", "example": "password123" }
        },
        "required": ["email", "password"]
      },
      "AuthResponse": {
        "type": "object",
        "description": "Response after successful user login",
        "properties": {
          "token": { "type": "string", "description": "JWT token", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." },
          "token_type": { "type": "string", "example": "cookie" },
          "expiry": { "type": "string", "example": "1 Hour" }
        },
        "required": ["token", "token_type", "expiry"]
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT token stored in an HttpOnly cookie"
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "Error": { "type": "boolean", "example": true },
                "Message": { "type": "string", "example": "Error in MYSQL query or server failure" }
              }
            }
          }
        }
      }
    }
  }
}